+++ src/walker.ts
@@ -57,6 +57,7 @@ export class GlobWalker {
   cache: GlobCache
   cwd: string
   start: string
+  parent?: GlobWalker
 
   constructor(
     pattern: Pattern,
@@ -75,6 +76,9 @@ export class GlobWalker {
       ignore,
       cwd = '',
     } = options
+    if (options instanceof GlobWalker) {
+      this.parent = options
+    }
 
     // if the pattern starts with a bunch of strings, then skip ahead
     this.pattern = [...pattern]
@@ -98,11 +102,14 @@ export class GlobWalker {
     }
   }
 
+  // TODO: this belongs in the Glob class probably
+  // since then we can remove the parent bit.
+  // Also, then we can ditch the cwd
   setStart() {
     const [first, ...rest] = this.pattern
     // a pattern like /a/s/d/f discards the cwd
     // a pattern like / (or c:/ on windows) can only match the root
-    if (typeof first === 'string') {
+    if (typeof first === 'string' && !this.parent) {
       const patternSlash = first === '' && !!rest.length
       const isWin = process.platform === 'win32'
       const patternDrive = isWin && /^[a-z]:$/i.test(first)
@@ -110,21 +117,53 @@ export class GlobWalker {
       if (setAbs) {
         // a pattern like '/' on windows goes to the root of
         // the drive that cwd is on.  If cwd isn't on a drive, use /
-        const cwd = this.join(this.path, this.cwd)
+        const cwd = this.cwd ? this.join(this.path, this.cwd) : this.path
         const cwdRe = /^[a-z]:($|[\\\/])/i
         const cwdDrive = isWin && patternSlash && cwd.match(cwdRe)
-        const root = cwdDrive ? cwdDrive[0] : '/'
+        // don't mount UNC paths on a drive letter
+        // eg: //?/c:/* or //host/share/*
+        // Only relevant if we WOULD have tried to mount it
+        // We'll gobble those strings shortly, so use '/' for now
+        const isUNC =
+          cwdDrive &&
+          first === '' &&
+          rest[0] === '' &&
+          rest[1] &&
+          typeof rest[1] === 'string' &&
+          rest[2] &&
+          typeof rest[2] === 'string'
+        const root = isUNC
+          ? '/'
+          : cwdDrive
+          ? cwdDrive[0]
+          : isWin
+          ? resolve('/')
+          : ''
+        if (isUNC && rest.length === 3) {
+          rest.push('')
+        }
         this.pattern = rest.length ? (rest as Pattern) : ['']
         this.cwd = first || root
-        this.path = first || root
+        this.path = first || root || '/'
       }
     }
     while (
       this.pattern.length > 1 &&
       typeof this.pattern[0] === 'string'
     ) {
-      this.path = this.join(this.pattern[0])
+      this.path =
+        this.path === '/'
+          ? this.path + this.pattern[0]
+          : this.join(this.pattern[0])
       this.pattern.shift()
+    }
+    while (
+      (this.pattern[0] instanceof RegExp ||
+        this.pattern[0] === GLOBSTAR) &&
+      this.pattern[1] === '' &&
+      this.pattern.length > 2
+    ) {
+      this.pattern.splice(1, 1)
     }
     return this.join(this.path, this.cwd) || '.'
   }
@@ -233,7 +272,7 @@ export class GlobWalker {
   join(p: string, base: string = this.path) {
     return base === ''
       ? p
-      : base.substring(base.length - 1) === '/'
+      : base === '/' && p
       ? `${base}${p}`
       : `${base}/${p}`
   }
@@ -244,7 +283,11 @@ export class GlobWalker {
     rest: Pattern | null
   ): Children {
     const children: Children = []
-    const e = entries.find(e => e.name === p)
+    const e =
+      p !== '' &&
+      p !== '.' &&
+      p !== '..' &&
+      entries.find(e => e.name === p)
 
     // since we pull all string portions up to minimize readdir() calls,
     // the only way we can possibly have a rest here is if something